% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_model.R
\name{avg_ps}
\alias{avg_ps}
\title{Compute average projection strengths within each cluster}
\usage{
avg_ps(Y, Z)
}
\arguments{
\item{Y}{a list of matrices. Each is a region-by-neuron matrix of projection counts for individual mouse.}

\item{Z}{a list of allocations (integers). Each is a vector of allocations for individual mouse.}
}
\value{
a matrix of projection strengths, with rows corresponding to clusters and columns for regions.
}
\description{
Compute average projection strengths within each cluster
}
\examples{
average_ps <- avg_ps(Y = Y, Z = Z)
}
