% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_model.R
\name{binomial_model}
\alias{binomial_model}
\title{Binomial Motif Analysis of MAPseq data}
\usage{
binomial_model(data, bin_thresh = 0, pval_cutoff = 0.05, binom_model = "two")
}
\arguments{
\item{data}{a list of matrices. Each is a region-by-neuron projection count matrix for an individual brain.}

\item{bin_thresh}{the binarization threshold. Barcode counts below threshold become 0. Barcode counts above threshold become 1.}

\item{pval_cutoff}{p-value threshold. Only motifs with p-values smaller than this threshold will be deemed "significant".}

\item{binom_model}{binomial model algorithm used. Options = c("one","two). Default = "two".}
}
\value{
results (list of motif analysis results)
}
\description{
Binomial Motif Analysis of MAPseq data
}
\examples{
res = binomial_model(list_of_matrices, bin_thresh = 5, pval_cutoff = 0.05, binom_model = "two")
}
