% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_model.R
\name{compare_brains}
\alias{compare_brains}
\title{Compare motif analysis results}
\usage{
compare_brains(results_1, results_2)
}
\arguments{
\item{results_1}{binomial model results object for Brain 1}

\item{results_2}{binomial model results object for Brain 2}
}
\value{
paired barchart comparing the estimated per-region projection probabilities of Brain 1 and Brain 2.
}
\description{
Compare motif analysis results
}
\examples{
compare_brains(brain1_results, brain2_results)
}
