% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_barseq}
\alias{data_barseq}
\title{The auditory cortex data}
\format{
A list of matrices. Each list contains a barcode-count matrix from a mouse, with rows for regions and columns for neurons.
}
\usage{
data_barseq
}
\description{
The auditory cortex data considers projections from the auditory cortex to the target regions:
the orbitofrontal cortex (OFC), motor cortex (Motor), rostral striatum (Rstr), somatosensory cortex (SSctx),
caudal striatum (Cstr), amygdala (Amyg), ipsilateral visual cortex (VisIp), contralateral visual cortex (VisC),
contralateral auditory cortex (AudC), thalamus (Thal), and tectum (Tect).
}
\references{
Chen, X., Sun, Y.-C., Zhan, H. et al. High-throughput mapping of long-range neuronal projection using in situ sequencing. Cell 179, 772-786 (2019).
}
\keyword{datasets}
