% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_Hans}
\alias{data_Hans}
\title{The visual cortex data}
\format{
A list of matrices for four mice. Each is a barcode-count matrix with rows for regions and columns for neurons.
}
\usage{
data_Hans
}
\description{
This data describes the projections of single neurons from the primary visual cortex to the to the
lateroitermediate (LI), lateromedial (LM), anterolateral (AL), posteromedial (PM), anteromedial (AM),
and rostrolateral (RL) higher visual areas.
}
\references{
Han, Y., Kebschull, J., Campbell, R. et al. The logic of single-cell projections from visual cortex. Nature 556, 51–56 (2018).
}
\keyword{datasets}
