% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data_generation.R
\name{data_simulation}
\alias{data_simulation}
\title{Title}
\usage{
data_simulation(
  mcmc_run_all_output,
  Y,
  regions.name = NULL,
  M,
  C,
  noise.levels = NULL
)
}
\arguments{
\item{noise.levels}{}
}
\description{
Title
}
