% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prominent_motif.R
\name{identify_prominent_motif}
\alias{identify_prominent_motif}
\title{Title}
\usage{
identify_prominent_motif(post_output_reorder, thresh = 0.02, prob = 0.95)
}
\arguments{
\item{prob}{}
}
\description{
Title
}
