% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_model.R
\name{k_means_axon}
\alias{k_means_axon}
\title{K-means clustering for MAPseq data}
\usage{
k_means_axon(Y, k, transformation = "row-sum", restart = 50, iter.max = 100)
}
\arguments{
\item{Y}{a list of matrices. Each is a region-by-neuron matrix of projection counts for individual mouse.}

\item{k}{number of clusters to be found.}

\item{transformation}{type of data transformation. Default to 'row-sum'. Can also be one of the following:
\itemize{
\item 'row-sum': normalize each neuron by its total counts.
\item 'log row-sum': after row-sum normalization, take log(y+1).
\item 'max-row': normalize each neuron by its maximum count.
\item 'cosine': normalize each neuron's counts to unit length.
}}

\item{restart}{number of restart in k-means algorithm.}

\item{iter.max}{the maximum number of iterations allowed in k-means.}
}
\value{
a list of allocations (integers). Each is a vector of allocations for individual mouse.
}
\description{
This function implements k-means clustering for MAPseq data with different data transformations.
}
\examples{
set.seed(43)
Z <- k_means_axon(Y = Y, k = 20, transformation = 'log row-sum', restart = 20, iter.max = 100)
}
