% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_reorder.R
\name{mcmc_reorder_cluster}
\alias{mcmc_reorder_cluster}
\title{Title}
\usage{
mcmc_reorder_cluster(
  post_output,
  Z,
  regions.name = NULL,
  epsilon = 0.01,
  cred.int = 0.9
)
}
\arguments{
\item{cred.int}{}
}
\description{
Title
}
