% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_size.R
\name{opt.clustering.frequency}
\alias{opt.clustering.frequency}
\title{Summarize the size of each cluster}
\usage{
opt.clustering.frequency(
  clustering,
  group.index = NULL,
  group.name = "group",
  cluster.index = NULL,
  title = "",
  col = NULL
)
}
\arguments{
\item{clustering}{a list of allocations (integers). Each is a vector of allocations for individual mouse.}

\item{group.index}{optional. A list of integers denoting the grouping, same dimension as \code{clustering}.}

\item{group.name}{a character for the title of the legend.}

\item{cluster.index}{optional. A vector of cluster labels (integers) to that will be plotted.}

\item{title}{the title of the barplot.}

\item{col}{optional. A vector of color names for the grouping.}
}
\value{
a barplot with cluster size given by the height of each bar. If grouping is provided, mice are colored by groups.
}
\description{
This function shows a barplot of cluster size, with the option to color by groups/mice.
}
