% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{plot_empirical_ps}
\alias{plot_empirical_ps}
\title{Title}
\usage{
plot_empirical_ps(
  Y,
  Z,
  cluster.labels = NULL,
  regions.name = NULL,
  group.index = NULL,
  group.name = "group",
  cluster.index = NULL,
  title = "",
  facet_ncol = 5,
  col = NULL
)
}
\arguments{
\item{col}{}
}
\description{
Title
}
