% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{plot_estimated_ps}
\alias{plot_estimated_ps}
\title{Title}
\usage{
plot_estimated_ps(
  ps_summary,
  cluster.index = NULL,
  title = "",
  facet_ncol = 5,
  col = NULL
)
}
\arguments{
\item{col}{}
}
\description{
Title
}
