% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot.R
\name{plot_summary}
\alias{plot_summary}
\title{Title}
\usage{
plot_summary(
  params,
  eps,
  prominent_motifs_prob,
  global_weight_thresh = 0.01,
  data.source.label = "Mouse",
  regions.name = NULL,
  col_bar,
  col_mat = c("white", "blue"),
  height = 6,
  width = 2,
  barwidth = 0.99,
  legend = TRUE,
  legend_x = 0.9,
  legend_y = 0.95
)
}
\arguments{
\item{legend_y}{}
}
\description{
Title
}
