% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_prob.R
\name{projection_probability}
\alias{projection_probability}
\title{Title}
\usage{
projection_probability(
  Y,
  post_output_reorder,
  allocation_prob,
  motif_indices,
  ncol = 5
)
}
\arguments{
\item{ncol}{}
}
\description{
Title
}
