% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_model.R
\name{upset_plot}
\alias{upset_plot}
\title{Generate Motif Analysis Upset Plot}
\usage{
upset_plot(cluster_summary)
}
\arguments{
\item{cluster_summary}{data.frame describing key features of all identified motifs}
}
\value{
Motif Analysis Upset Plot
}
\description{
Generate Motif Analysis Upset Plot
}
\examples{
chart = upset_plot(res$cluster_summary)
}
