% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_Hans}
\alias{data_Hans}
\title{The visual cortex data}
\format{
A list of matrices. Each list contains a barcode-count matrix from a mouse, with rows for regions and columns for neurons.
}
\usage{
data_Hans
}
\description{
The visual cortex data describes the projections of single neurons from the primary visual cortex to the
laterointermediate (LI), lateromedial (LM), anterolateral (AL), posteromedial (PM), anteromedial (AM), and rostrolateral (RL) areas.
}
\references{
Han, Y., Kebschull, J.M., Campbell, R.A. et al. The logic of single-cell projections from visual cortex. Nature 556, 51–56 (2018).
}
\keyword{datasets}
