% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPMAP_mcmc.R
\name{HBMAP_mcmc}
\alias{HBMAP_mcmc}
\title{Title}
\usage{
HBMAP_mcmc(
  Y = NULL,
  J = NULL,
  mcmc = list(),
  prior = list(),
  Z.fix = NULL,
  post = NULL,
  Z.init = NULL,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
