% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_ps.R
\name{heatmap_ps}
\alias{heatmap_ps}
\title{Title}
\usage{
heatmap_ps(
  Y,
  Z,
  regions.name = NULL,
  group.index = NULL,
  group.name = "group",
  cluster.index = NULL,
  col = NULL,
  title = ""
)
}
\arguments{
\item{title}{}
}
\description{
Title
}
