% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_analysis.R
\name{local_weights_analysis}
\alias{local_weights_analysis}
\title{Title}
\usage{
local_weights_analysis(
  N = 100,
  Z,
  omega_output,
  omega_J_M_output,
  prior,
  mcmc,
  post,
  mouse_indices = NULL,
  n_cores = NULL,
  cluster_type = "FORK",
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
