% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_size.R
\name{opt.clustering.frequency}
\alias{opt.clustering.frequency}
\title{Title}
\usage{
opt.clustering.frequency(
  clustering,
  group.index = NULL,
  group.name = "group",
  cluster.index = NULL,
  title = "",
  col = NULL
)
}
\arguments{
\item{col}{}
}
\description{
Title
}
