% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_similarity_plot.R
\name{plotpsm}
\alias{plotpsm}
\title{Plot posterior similarity matrix}
\usage{
plotpsm(psm.ind, psm.tot, method = "complete", title = "")
}
\arguments{
\item{psm.ind}{a list of posterior similarity matrices between pairwise mice. The first output from \code{similarity_matrix}.}

\item{psm.tot}{a single posterior similarity for all mice together. The second output from \code{similarity_matrix}.}

\item{method}{the method to perform hierarchical clustering to arrange posterior similarity matrix. see \code{hclust}.}

\item{title}{the title of the first plot.}
}
\value{
two ggplots
\item{plot.tot}{a heatmap of posterior similarity matrix, without separating mice.}
\item{plot.ind}{a heatmap of posterior similarity matrix, with mice separated by black solid lines, the main diagonal blocks correspond to within-mouse similarity.}
}
\description{
Plot posterior similarity matrix
}
